+++ src/store/actionTypes.ts
@@ -1,0 +1,23 @@
+/**
+ * 用户相关
+ */
+export const USER = {
+  /**
+   * 设置用户信息
+   */
+  USER_SET_USER_INFO: 'USER_SET_USER_INFO',
+  /**
+   * 重置用户信息
+   */
+  USER_RESET_USER_INFO: 'USER_RESET_USER_INFO',
+};
+
+/**
+ * app 相关
+ */
+export const APP = {
+  /**
+   * 设置APP的侧滑栏状态
+   */
+  APP_TOGGLE_SIDEBAR: 'APP_TOGGLE_SIDEBAR',
+};