+++ src/store/reducers/modules/user.ts
@@ -1,0 +1,6 @@
+import { IAction } from '@/types/store';
+import { IUserState } from '@/types/store';
+
+export default function userReducer(state: IUserState, action: IAction) {
+  return state;
+}