+++ src/types/index.ts
@@ -1,0 +1,6 @@
+/**
+ * 通用对象
+ */
+export interface IObject {
+  [propName: string]: any;
+}